proposing changes to the project
Due to the blockade encountered for the case-study in the previous proposal, the changes are proposed to execute the project without committing to the same case study. However, the replacement case study is currently being explored and will be communicated soon with the mentor and organizers. Meanwhile, please consider these changes to my proposal.